// client/modules/impower-script-engine/types/CommandType.ts
export type CommandType =
  | "EvalStart"
  | "EvalOutput"
  | "EvalEnd"
  | "Duplicate"
  | "PopEvaluatedValue"
  | "PopFunction"
  | "PopTunnel"
  | "BeginString"
  | "EndString"
  | "NoOp"
  | "ChoiceCount"
  | "Turns"
  | "TurnsSince"
  | "Random"
  | "SeedRandom"
  | "VisitIndex"
  | "SequenceShuffleIndex"
  | "StartThread"
  | "Done"
  | "End"
  | "ListFromInt"
  | "ListRange"
  | "ListRandom"
  | "ReadCount";